Finalized framework of testing loop

+++ 3D Molecular Computation Package/src/simulation/TimeUpdatable.java
@@ -1,5 +1,5 @@
 package simulation;
 
 public interface TimeUpdatable extends Updatable{
-	public void update(long milliseconds);
+	public void update(double milliseconds);
 }
